// src/social-media/enums/socialMediaName.enum.ts
export enum SocialMediaName {
  INSTAGRAM = 'instagram',
  FACEBOOK = 'facebook',
  TWITTER = 'twitter',
  YOUTUBE = 'youtube',
  LINKEDIN = 'linkedin',
  TIKTOK = 'tiktok',
  SNAPCHAT = 'snapchat',
  WHATSAPP = 'whatsapp',
  DISCORD = 'discord',
  SPOTIFY = 'spotify',
  SOUNDCLOUD = 'soundcloud',
}